# fix: fix docarray iter