@davidbp 2021-06-10
I did some benchmarking on the updated `__iter__` method and it does not seem to degrade at all.

I used the following test
```
n_samples = 5

def iterate_over_d(darr):
    i = 0
    for d in darr:
        i+=1

for i in range(5):
    n_samples = n_samples *10
    darr = DocumentArray([jina_pb2.DocumentProto()]*n_samples)
    print(f'n_samples={n_samples}')
    %timeit iterate_over_d(darr)
```


#### This branch results

```
n_samples=50
393 µs ± 22.1 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
n_samples=500
3.74 ms ± 60.1 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
n_samples=5000
36.7 ms ± 285 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
n_samples=50000
369 ms ± 4.13 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
n_samples=500000
3.74 s ± 43.1 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

#### Master branch results

```
n_samples=50
369 µs ± 4.39 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
n_samples=500
3.68 ms ± 59.1 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
n_samples=5000
37.7 ms ± 686 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
n_samples=50000
381 ms ± 15.7 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
n_samples=500000
3.74 s ± 72.1 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```